be: refactoring - setup dirs
