--- repository/orders/__init__.py
@@ -1,3 +1,0 @@
-from .Order import OrderRepository
-
-__all__ = ["OrderRepository"]
