+++ schemas/booking/__init__.py
@@ -1,2 +1,3 @@
 from .Dish import DishCreate, DishRead, DishUpdate, DishFilter
 from .OrderItem import OrderItemCreate, OrderItemRead, OrderItemUpdate, OrderItemFilter
+from .Order import OrderCreate, OrderRead, OrderUpdate, OrderFilter
