--- schemas/orders/__init__.py
@@ -1,13 +1,0 @@
-from .Order import (
-    OrderCreate,
-    OrderUpdate,
-    OrderRead,
-    OrderFilter,
-)
-
-__all__ = [
-    "OrderCreate",
-    "OrderUpdate",
-    "OrderRead",
-    "OrderFilter",
-]
